[PAA-1336] Removed backup files created druring update to VS2015.
