--- Backup/DataFresh/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-using System.Reflection;
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyTitleAttribute("dataFresh")]
-[assembly: DataFresh.AssemblyResourceEncrpytion(false)]
