--- Backup/DataFreshUtil/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-using System.Reflection;
-
-[assembly: AssemblyVersionAttribute("1.0.0.0")]
-[assembly: AssemblyTitleAttribute("dataFresh Console Utility")]
-
