--- Backup/DataFreshUtil/EntryPoint.cs
@@ -1,31 +1,0 @@
-// EntropyZero dataFresh Copyright (C) 2007 EntropyZero Consulting, LLC.
-// Please visit us on the web: http://blogs.ent0.com/
-//
-// This library is free software; you can redistribute it and/or modify 
-// it under the terms of the GNU Lesser General Public License as 
-// published by the Free Software Foundation; either version 2.1 of the 
-// License, or (at your option) any later version.
-//
-// This library is distributed in the hope that it will be useful, but 
-// WITHOUT ANY WARRANTY; without even the implied warranty of 
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
-// Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public 
-// License along with this library; if not, write to:
-// Free Software Foundation, Inc., 
-// 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA 
-
-using DataFresh;
-
-namespace DataFreshUtil
-{
-	public class EntryPoint
-	{
-		public static void Main(string[] args)
-		{
-			DataFreshConsole console = new DataFreshConsole();
-			console.Start(args);
-		}
-	}
-}
