--- Backup/EntropyZero DataFresh Solution.sln
@@ -1,51 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 8.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataFresh", "DataFresh\DataFresh.csproj", "{05438358-2120-4D97-974E-ACF80F7EA6F7}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestHarnessExample", "TestHarnessExample\TestHarnessExample.csproj", "{D688D8F0-3CDE-4C39-BF81-ED8C210C8F8E}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Testing.DataFresh", "Testing.DataFresh\Testing.DataFresh.csproj", "{3E910EED-904F-4A52-A611-9278B682EFD3}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataFreshUtil", "DataFreshUtil\DataFreshUtil.csproj", "{D1044AB1-AB58-42E3-A37B-4099116982A5}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataFresh.Common", "DataFresh.Common\DataFresh.Common.csproj", "{0BAA8852-1597-43BB-AF78-BABA707D4D86}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfiguration) = preSolution
-		Debug = Debug
-		Release = Release
-	EndGlobalSection
-	GlobalSection(ProjectConfiguration) = postSolution
-		{05438358-2120-4D97-974E-ACF80F7EA6F7}.Debug.ActiveCfg = Debug|.NET
-		{05438358-2120-4D97-974E-ACF80F7EA6F7}.Debug.Build.0 = Debug|.NET
-		{05438358-2120-4D97-974E-ACF80F7EA6F7}.Release.ActiveCfg = Release|.NET
-		{05438358-2120-4D97-974E-ACF80F7EA6F7}.Release.Build.0 = Release|.NET
-		{D688D8F0-3CDE-4C39-BF81-ED8C210C8F8E}.Debug.ActiveCfg = Debug|.NET
-		{D688D8F0-3CDE-4C39-BF81-ED8C210C8F8E}.Debug.Build.0 = Debug|.NET
-		{D688D8F0-3CDE-4C39-BF81-ED8C210C8F8E}.Release.ActiveCfg = Release|.NET
-		{D688D8F0-3CDE-4C39-BF81-ED8C210C8F8E}.Release.Build.0 = Release|.NET
-		{3E910EED-904F-4A52-A611-9278B682EFD3}.Debug.ActiveCfg = Debug|.NET
-		{3E910EED-904F-4A52-A611-9278B682EFD3}.Debug.Build.0 = Debug|.NET
-		{3E910EED-904F-4A52-A611-9278B682EFD3}.Release.ActiveCfg = Release|.NET
-		{3E910EED-904F-4A52-A611-9278B682EFD3}.Release.Build.0 = Release|.NET
-		{D1044AB1-AB58-42E3-A37B-4099116982A5}.Debug.ActiveCfg = Debug|.NET
-		{D1044AB1-AB58-42E3-A37B-4099116982A5}.Debug.Build.0 = Debug|.NET
-		{D1044AB1-AB58-42E3-A37B-4099116982A5}.Release.ActiveCfg = Release|.NET
-		{D1044AB1-AB58-42E3-A37B-4099116982A5}.Release.Build.0 = Release|.NET
-		{0BAA8852-1597-43BB-AF78-BABA707D4D86}.Debug.ActiveCfg = Debug|.NET
-		{0BAA8852-1597-43BB-AF78-BABA707D4D86}.Release.ActiveCfg = Release|.NET
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-	EndGlobalSection
-	GlobalSection(ExtensibilityAddIns) = postSolution
-	EndGlobalSection
-EndGlobal
