--- Backup/TestHarnessExample/Readme.txt
@@ -1,11 +1,0 @@
-Using the dataFresh Sample
-
-1. Copy the files from the Scripts folder to a folder on your desktop.
-
-2. Copy and DataFreshUtil.exe to the same folder on your desktop.
-
-3. Create a database on your local sql server instance names DataFreshSample
-
-4. Modify the appropriate authentication arguments to login to your sql server.
-
-5. Execute "Create Database.bat" -- This will populate the DataFreshSample database and prepare it by adding the DataFresh elements.
