--- Backup/TestHarnessExample/Scripts/Create Database.bat
@@ -1,5 +1,0 @@
-osql.exe -S 127.0.0.1 -U test -P test -d DataFreshSample -i database.sql
-
-DataFreshUtil.exe -s 127.0.0.1 -u test -p test -d DataFreshSample -c PREPARE
-
-pause
