--- Backup/TestHarnessExample/Scripts/DataFresh Usage.bat
@@ -1,2 +1,0 @@
-DataFreshUtil.exe
-pause
