--- Backup/TestHarnessExample/Scripts/Refresh Database.bat
@@ -1,3 +1,0 @@
-DataFreshUtil.exe -s 127.0.0.1 -u test -p test -d DataFreshSample -c REFRESH
-
-pause
